CF use LogAwsUserInfoConvention instead
